app/requirements.txt
fastapi
uvicorn
apscheduler
requests
bs4
lxml
xmltodict
sqlalchemy
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cymem==2.0.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.1
pytz==2024.2
regex==2024.9.11
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
starlette==0.38.6
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wasabi==1.1.3
wrapt==1.16.0
firebase-admin
